Fixup size_t

+++ src/NumericalAlgorithms/SphericalHarmonics/TensorYlmFilter.cpp
@@ -8,10 +8,10 @@
 #include <optional>
 
 #include "DataStructures/SparseMatrixFiller.hpp"
+#include "DataStructures/Tensor/TypeAliases.hpp"
 #include "NumericalAlgorithms/SphericalHarmonics/SpherepackIterator.hpp"
 #include "NumericalAlgorithms/SphericalHarmonics/TensorYlmHelpers.hpp"
 #include "NumericalAlgorithms/SphericalHarmonics/WignerThreeJ.hpp"
-#include "DataStructures/Tensor/TypeAliases.hpp"
 
 namespace ylm::TensorYlm {
 
@@ -41,46 +41,62 @@ void inner_loops_one(SparseMatrixFiller& filler, SpherepackIterator& iter_src,
       if (msrc > 0) {
         // Main term.
         // ReRe
-        iter_src.set(ell, msrc, SpherepackIterator::CoefficientArray::a);
-        iter_dest.set(ell, mdest, SpherepackIterator::CoefficientArray::a);
+        iter_src.set(ell, static_cast<size_t>(msrc),
+                     SpherepackIterator::CoefficientArray::a);
+        iter_dest.set(ell, static_cast<size_t>(mdest),
+                      SpherepackIterator::CoefficientArray::a);
         add_element(correction.real());
 
         // ReIm
-        iter_src.set(ell, msrc, SpherepackIterator::CoefficientArray::b);
-        iter_dest.set(ell, mdest, SpherepackIterator::CoefficientArray::a);
+        iter_src.set(ell, static_cast<size_t>(msrc),
+                     SpherepackIterator::CoefficientArray::b);
+        iter_dest.set(ell, static_cast<size_t>(mdest),
+                      SpherepackIterator::CoefficientArray::a);
         add_element(-correction.imag());
 
         // ImIm
-        iter_src.set(ell, msrc, SpherepackIterator::CoefficientArray::b);
-        iter_dest.set(ell, mdest, SpherepackIterator::CoefficientArray::b);
+        iter_src.set(ell, static_cast<size_t>(msrc),
+                     SpherepackIterator::CoefficientArray::b);
+        iter_dest.set(ell, static_cast<size_t>(mdest),
+                      SpherepackIterator::CoefficientArray::b);
         add_element(correction.real());
 
         // ImRe
-        iter_src.set(ell, msrc, SpherepackIterator::CoefficientArray::a);
-        iter_dest.set(ell, mdest, SpherepackIterator::CoefficientArray::b);
+        iter_src.set(ell, static_cast<size_t>(msrc),
+                     SpherepackIterator::CoefficientArray::a);
+        iter_dest.set(ell, static_cast<size_t>(mdest),
+                      SpherepackIterator::CoefficientArray::b);
         add_element(correction.imag());
       } else {
         // We are multiplying by Tlmsrc but we should be
         // multiplying by (Tlmsrc)^star (-1)^msrc
         const double sign = (msrc % 2 == 0 ? 1.0 : -1.0);
         // ReRe
-        iter_src.set(ell, -msrc, SpherepackIterator::CoefficientArray::a);
-        iter_dest.set(ell, mdest, SpherepackIterator::CoefficientArray::a);
+        iter_src.set(ell, static_cast<size_t>(-msrc),
+                     SpherepackIterator::CoefficientArray::a);
+        iter_dest.set(ell, static_cast<size_t>(mdest),
+                      SpherepackIterator::CoefficientArray::a);
         add_element(sign * correction.real());
 
         // ReIm
-        iter_src.set(ell, -msrc, SpherepackIterator::CoefficientArray::b);
-        iter_dest.set(ell, mdest, SpherepackIterator::CoefficientArray::a);
+        iter_src.set(ell, static_cast<size_t>(-msrc),
+                     SpherepackIterator::CoefficientArray::b);
+        iter_dest.set(ell, static_cast<size_t>(mdest),
+                      SpherepackIterator::CoefficientArray::a);
         add_element(sign * correction.imag());
 
         // ImRe
-        iter_src.set(ell, -msrc, SpherepackIterator::CoefficientArray::a);
-        iter_dest.set(ell, mdest, SpherepackIterator::CoefficientArray::b);
+        iter_src.set(ell, static_cast<size_t>(-msrc),
+                     SpherepackIterator::CoefficientArray::a);
+        iter_dest.set(ell, static_cast<size_t>(mdest),
+                      SpherepackIterator::CoefficientArray::b);
         add_element(sign * correction.imag());
 
         // ImIm
-        iter_src.set(ell, -msrc, SpherepackIterator::CoefficientArray::b);
-        iter_dest.set(ell, mdest, SpherepackIterator::CoefficientArray::b);
+        iter_src.set(ell, static_cast<size_t>(-msrc),
+                     SpherepackIterator::CoefficientArray::b);
+        iter_dest.set(ell, static_cast<size_t>(mdest),
+                      SpherepackIterator::CoefficientArray::b);
         add_element(-sign * correction.real());
       }
     }
@@ -147,59 +163,59 @@ void inner_loops_two(SparseMatrixFiller& filler, SpherepackIterator& iter_src,
                 if (m_src > 0) {
                   // Main term.
                   // ReRe
-                  iter_src.set(l_dest, m_src,
+                  iter_src.set(l_dest, static_cast<size_t>(m_src),
                                SpherepackIterator::CoefficientArray::a);
-                  iter_dest.set(l_dest, m_dest,
+                  iter_dest.set(l_dest, static_cast<size_t>(m_dest),
                                 SpherepackIterator::CoefficientArray::a);
                   add_element(correction.real());
 
                   // ReIm
-                  iter_src.set(l_dest, m_src,
+                  iter_src.set(l_dest, static_cast<size_t>(m_src),
                                SpherepackIterator::CoefficientArray::b);
-                  iter_dest.set(l_dest, m_dest,
+                  iter_dest.set(l_dest, static_cast<size_t>(m_dest),
                                 SpherepackIterator::CoefficientArray::a);
                   add_element(-correction.imag());
 
                   // ImIm
-                  iter_src.set(l_dest, m_src,
+                  iter_src.set(l_dest, static_cast<size_t>(m_src),
                                SpherepackIterator::CoefficientArray::b);
-                  iter_dest.set(l_dest, m_dest,
+                  iter_dest.set(l_dest, static_cast<size_t>(m_dest),
                                 SpherepackIterator::CoefficientArray::b);
                   add_element(correction.real());
 
                   // ImRe
-                  iter_src.set(l_dest, m_src,
+                  iter_src.set(l_dest, static_cast<size_t>(m_src),
                                SpherepackIterator::CoefficientArray::a);
-                  iter_dest.set(l_dest, m_dest,
+                  iter_dest.set(l_dest, static_cast<size_t>(m_dest),
                                 SpherepackIterator::CoefficientArray::b);
                   add_element(correction.imag());
                 } else {
                   const double sign = (m_src % 2 == 0 ? 1.0 : -1.0);
                   // ReRe
-                  iter_src.set(l_dest, -m_src,
+                  iter_src.set(l_dest, static_cast<size_t>(-m_src),
                                SpherepackIterator::CoefficientArray::a);
-                  iter_dest.set(l_dest, m_dest,
+                  iter_dest.set(l_dest, static_cast<size_t>(m_dest),
                                 SpherepackIterator::CoefficientArray::a);
                   add_element(sign * correction.real());
 
                   // ReIm
-                  iter_src.set(l_dest, -m_src,
+                  iter_src.set(l_dest, static_cast<size_t>(-m_src),
                                SpherepackIterator::CoefficientArray::b);
-                  iter_dest.set(l_dest, m_dest,
+                  iter_dest.set(l_dest, static_cast<size_t>(m_dest),
                                 SpherepackIterator::CoefficientArray::a);
                   add_element(sign * correction.imag());
 
                   // ImRe
-                  iter_src.set(l_dest, -m_src,
+                  iter_src.set(l_dest, static_cast<size_t>(-m_src),
                                SpherepackIterator::CoefficientArray::a);
-                  iter_dest.set(l_dest, m_dest,
+                  iter_dest.set(l_dest, static_cast<size_t>(m_dest),
                                 SpherepackIterator::CoefficientArray::b);
                   add_element(sign * correction.imag());
 
                   // ImIm
-                  iter_src.set(l_dest, -m_src,
+                  iter_src.set(l_dest, static_cast<size_t>(-m_src),
                                SpherepackIterator::CoefficientArray::b);
-                  iter_dest.set(l_dest, m_dest,
+                  iter_dest.set(l_dest, static_cast<size_t>(m_dest),
                                 SpherepackIterator::CoefficientArray::b);
                   add_element(-sign * correction.real());
                 }
@@ -302,47 +318,47 @@ void inner_loops_three(
                 if (m_src > 0) {
                   // Main term.
                   // ReRe
-                  iter_src.set(l_dest, m_src,
+                  iter_src.set(l_dest, static_cast<size_t>(m_src),
                                SpherepackIterator::CoefficientArray::a);
-                  iter_dest.set(l_dest, m_dest,
+                  iter_dest.set(l_dest, static_cast<size_t>(m_dest),
                                 SpherepackIterator::CoefficientArray::a);
                   add_element(correction.real());
 
                   // ReIm
-                  iter_src.set(l_dest, m_src,
+                  iter_src.set(l_dest, static_cast<size_t>(m_src),
                                SpherepackIterator::CoefficientArray::b);
                   add_element(-correction.imag());
 
                   // ImIm
-                  iter_dest.set(l_dest, m_dest,
+                  iter_dest.set(l_dest, static_cast<size_t>(m_dest),
                                 SpherepackIterator::CoefficientArray::b);
                   add_element(correction.real());
 
                   // ImRe
-                  iter_src.set(l_dest, m_src,
+                  iter_src.set(l_dest, static_cast<size_t>(m_src),
                                SpherepackIterator::CoefficientArray::a);
                   add_element(correction.imag());
                 } else {
                   const double sign = (m_src % 2 == 0 ? 1.0 : -1.0);
                   // ReRe
-                  iter_src.set(l_dest, -m_src,
+                  iter_src.set(l_dest, static_cast<size_t>(-m_src),
                                SpherepackIterator::CoefficientArray::a);
-                  iter_dest.set(l_dest, m_dest,
+                  iter_dest.set(l_dest, static_cast<size_t>(m_dest),
                                 SpherepackIterator::CoefficientArray::a);
                   add_element(sign * correction.real());
 
                   // ReIm
-                  iter_src.set(l_dest, -m_src,
+                  iter_src.set(l_dest, static_cast<size_t>(-m_src),
                                SpherepackIterator::CoefficientArray::b);
                   add_element(sign * correction.imag());
 
                   // ImIm
-                  iter_dest.set(l_dest, m_dest,
+                  iter_dest.set(l_dest, static_cast<size_t>(m_dest),
                                 SpherepackIterator::CoefficientArray::b);
                   add_element(-sign * correction.real());
 
                   // ImRe
-                  iter_src.set(l_dest, -m_src,
+                  iter_src.set(l_dest, static_cast<size_t>(-m_src),
                                SpherepackIterator::CoefficientArray::a);
                   add_element(sign * correction.imag());
                 }
